requirements.txt
Django==1.10.4
Pillow==2.6.2
Unipath==1.1
django-extensions==1.5.5
requests==2.7.0
python-decouple==2.3
pytz==2015.4
beautifulsoup4==4.3.2
xlwt==1.1.2
django-ajaximage
